+++ cms/src/layouts/ProtectedLayout.tsx
@@ -10,7 +10,6 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
   console.log("ProtectedLayout running");
   const { accessToken, setAccessToken } = useAuth();
   const location = useLocation();
-  // const {prevLocation} = usePrevLocation();
   // console.log(prevLocation);
   // if (!accessToken)
   //   return <Navigate to="/login" state={{ from: location }} replace />;
@@ -23,21 +22,39 @@ const ProtectedLayout: React.FC<{ children?: React.ReactNode }> = ({
     //   /* if (!accessToken)
     //     navigate("/login", { state: { from: location }, replace: true }); */
     const authorize = async () => {
-      try {
-        console.log("accessToken in authorize():", accessToken);
+      /* if (accessToken) {
+        try {
+          console.log("accessToken in authorize():", accessToken);
 
-        if (accessToken)
           axios.defaults.headers.common[
             "Authorization"
           ] = `Bearer ${accessToken}`;
+          const response = await axios.post("/auth/refresh");
+
+          console.log("response:", response);
+          setAccessToken(response.data.data.accessToken);
+        } catch (err) {
+          console.log(err);
+          setAccessToken(null);
+          navigate("/login");
+        }
+      } */
+
+      try {
+        console.log("accessToken in authorize():", accessToken);
+
+        axios.defaults.headers.common[
+          "Authorization"
+        ] = `Bearer ${accessToken}`;
         const response = await axios.post("/auth/refresh");
 
         console.log("response:", response);
         setAccessToken(response.data.data.accessToken);
       } catch (err) {
+        console.log(location);
         console.log(err);
         setAccessToken(null);
-        navigate("/login");
+        navigate("/login", { state: { prevLocation: location.pathname } });
       }
     };